1358 arcade keycloak rbac (#6)
* working user role actions

* Refactored out role type and role enums, structure improvements

* Improved and fixed server admin role parsing and applied to server actions

* revert debug test changes

* arcade auto creates and assigns new role for newly created database

* Removed OBE debugging

* Additional cleanup

* Doc improvements, removed OBE debugging code

* moar cleanup

* Moar cleanup

* more stuff

* type based access control appears to work

* fixed multiple groups worth of dba perms overwriting each other

* added good comment

* cleanup

* removed obe deps

* additional cleanup

* Fixed User group roles not being included in data access enforcement decisions

* Fixed server admin command authorization, post database role assigning

* cleaned up obe logging

+++ server/src/main/java/com/arcadedb/server/http/handler/PostServerCommandHandler.java
@@ -26,7 +26,6 @@ import com.arcadedb.engine.PaginatedFile;
 import com.arcadedb.exception.CommandExecutionException;
 import com.arcadedb.network.binary.ServerIsNotTheLeaderException;
 import com.arcadedb.serializer.json.JSONArray;
-import com.arcadedb.serializer.json.JSONObject;
 import com.arcadedb.server.ArcadeDBServer;
 import com.arcadedb.server.ha.HAServer;
 import com.arcadedb.server.ha.Leader2ReplicaNetworkExecutor;
@@ -36,6 +35,15 @@ import com.arcadedb.server.ha.message.ServerShutdownRequest;
 import com.arcadedb.server.http.HttpServer;
 import com.arcadedb.server.security.ServerSecurityException;
 import com.arcadedb.server.security.ServerSecurityUser;
+import com.arcadedb.server.security.oidc.ArcadeRole;
+import com.arcadedb.server.security.oidc.KeycloakClient;
+import com.arcadedb.server.security.oidc.role.CRUDPermission;
+import com.arcadedb.server.security.oidc.role.DatabaseAdminRole;
+import com.arcadedb.server.security.oidc.role.RoleType;
+import com.arcadedb.server.security.oidc.role.ServerAdminRole;
+import com.google.gson.JsonParser;
+import com.nimbusds.oauth2.sdk.util.StringUtils;
+
 import io.undertow.server.HttpServerExchange;
 
 import java.io.*;
@@ -53,48 +61,64 @@ public class PostServerCommandHandler extends AbstractHandler {
   }
 
   @Override
-  public ExecutionResponse execute(final HttpServerExchange exchange, final ServerSecurityUser user) throws IOException {
-    final JSONObject payload = new JSONObject(parseRequestPayload(exchange));
+  public ExecutionResponse execute(final HttpServerExchange exchange, final ServerSecurityUser user)
+      throws IOException {
+    String payloadString = parseRequestPayload(exchange);
 
-    final String command = payload.has("command") ? payload.getString("command") : null;
-    if (command == null)
-      return new ExecutionResponse(400, "{ \"error\" : \"Server command is null\"}");
+    if (payloadString != null && StringUtils.isNotBlank(payloadString)) {
+      final var payload = JsonParser.parseString(payloadString).getAsJsonObject();
 
-    if (!command.equals("list databases"))
-      checkRootUser(user);
+      final String command = payload.has("command") ? payload.get("command").getAsString() : null;
+      if (command == null)
+        return new ExecutionResponse(400, "{ \"error\" : \"Server command is null\"}");
 
-    if (command.startsWith("shutdown"))
-      shutdownServer(command);
-    else if (command.startsWith("create database "))
-      createDatabase(command);
-    else if (command.equals("list databases")) {
-      return listDatabases(user);
-    } else if (command.startsWith("drop database "))
-      dropDatabase(command);
-    else if (command.startsWith("close database "))
-      closeDatabase(command);
-    else if (command.startsWith("open database "))
-      openDatabase(command);
-    else if (command.startsWith("create user "))
-      createUser(command);
-    else if (command.startsWith("drop user "))
-      dropUser(command);
-    else if (command.startsWith("connect cluster ")) {
-      if (!connectCluster(command, exchange))
-        return null;
-    } else if (command.equals("disconnect cluster"))
-      disconnectCluster();
-    else if (command.startsWith("set database setting "))
-      setDatabaseSetting(command);
-    else if (command.startsWith("set server setting "))
-      setServerSetting(command);
-    else if (command.startsWith("get server events"))
-      return getServerEvents(command);
-    else if (command.startsWith("align database "))
-      alignDatabase(command);
-    else {
-      httpServer.getServer().getServerMetrics().meter("http.server-command.invalid").hit();
-      return new ExecutionResponse(400, "{ \"error\" : \"Server command not valid\"}");
+      // List of server commands that will check their own user permissions
+      var excludedTopLevelCheckComamnds = List.of("list databases", "create database", "drop database");
+
+      var isExcludedCommand = excludedTopLevelCheckComamnds.stream()
+          .anyMatch(excludedCommand -> command.startsWith(excludedCommand));
+
+      // If not a command that manages its own permissions, check if user has sa role
+      if (!isExcludedCommand
+          && !httpServer.getServer().getSecurity().checkUserHasAnyServerAdminRole(user, List.of(ServerAdminRole.ALL))) {
+        throw new ServerSecurityException(
+            String.format("User '%s' is not authorized to execute server command '%s'", user.getName(), command));
+      }
+
+      if (command.startsWith("shutdown"))
+        shutdownServer(command);
+      else if (command.startsWith("create database "))
+        createDatabase(command, user);
+      else if (command.equals("list databases")) {
+        return listDatabases(user);
+      } else if (command.startsWith("drop database "))
+        dropDatabase(command, user);
+      else if (command.startsWith("close database "))
+        closeDatabase(command);
+      else if (command.startsWith("open database "))
+        openDatabase(command);
+      else if (command.startsWith("create user ")) {
+        // createUser(command);
+        return new ExecutionResponse(400, "{ \"error\" : \"Please create new users through keycloak.\"}");
+      } else if (command.startsWith("drop user "))
+        dropUser(command);
+      else if (command.startsWith("connect cluster ")) {
+        if (!connectCluster(command, exchange))
+          return null;
+      } else if (command.equals("disconnect cluster"))
+        disconnectCluster();
+      else if (command.startsWith("set database setting "))
+        setDatabaseSetting(command);
+      else if (command.startsWith("set server setting "))
+        setServerSetting(command);
+      else if (command.startsWith("get server events"))
+        return getServerEvents(command);
+      else if (command.startsWith("align database "))
+        alignDatabase(command);
+      else {
+        httpServer.getServer().getServerMetrics().meter("http.server-command.invalid").hit();
+        return new ExecutionResponse(400, "{ \"error\" : \"Server command not valid\"}");
+      }
     }
 
     return new ExecutionResponse(200, "{ \"result\" : \"ok\"}");
@@ -112,6 +136,8 @@ public class PostServerCommandHandler extends AbstractHandler {
   }
 
   private void setServerSetting(final String command) {
+    // TODO check if user has sa role, or is root
+
     final String pair = command.substring("set server setting ".length());
     final String[] keyValue = pair.split(" ");
     if (keyValue.length != 2)
@@ -163,20 +189,40 @@ public class PostServerCommandHandler extends AbstractHandler {
     });
   }
 
-  private void createDatabase(final String command) {
+  private void createDatabase(final String command, final ServerSecurityUser user) {
     final String databaseName = command.substring("create database ".length()).trim();
     if (databaseName.isEmpty())
       throw new IllegalArgumentException("Database name empty");
 
-    checkServerIsLeaderIfInHA();
+    // check if user has create database role, or is root
+    var anyRequiredRoles = List.of(ServerAdminRole.CREATE_DATABASE, ServerAdminRole.ALL);
+    if (httpServer.getServer().getSecurity().checkUserHasAnyServerAdminRole(user, anyRequiredRoles)) {
+      checkServerIsLeaderIfInHA();
 
-    final ArcadeDBServer server = httpServer.getServer();
-    server.getServerMetrics().meter("http.create-database").hit();
+      final ArcadeDBServer server = httpServer.getServer();
+      server.getServerMetrics().meter("http.create-database").hit();
 
-    final DatabaseInternal db = server.createDatabase(databaseName, PaginatedFile.MODE.READ_WRITE);
+      final DatabaseInternal db = server.createDatabase(databaseName, PaginatedFile.MODE.READ_WRITE);
 
-    if (server.getConfiguration().getValueAsBoolean(GlobalConfiguration.HA_ENABLED))
-      ((ReplicatedDatabase) db).createInReplicas();
+      if (server.getConfiguration().getValueAsBoolean(GlobalConfiguration.HA_ENABLED))
+        ((ReplicatedDatabase) db).createInReplicas();
+
+      // Create and assign new role granting all data, schema, and settings on the new
+      // database to the user who created it.
+      ArcadeRole dataRole = new ArcadeRole(RoleType.USER, databaseName, "*", CRUDPermission.getAll());
+      ArcadeRole schemaRole = new ArcadeRole(RoleType.DATABASE_ADMIN, databaseName, DatabaseAdminRole.ALL);
+      createAndAssignRoleToUser(dataRole, user.getName());
+      createAndAssignRoleToUser(schemaRole, user.getName());
+    } else {
+      throw new ServerSecurityException("Create database operation not allowed for user " + user.getName());
+    }
+  }
+
+  private void createAndAssignRoleToUser(ArcadeRole arcadeRole, String username) {
+    String newRole = arcadeRole.getKeycloakRoleName();
+    KeycloakClient.createRole(newRole);
+    KeycloakClient.assignRoleToUser(newRole, username);
+    httpServer.getServer().getSecurity().appendArcadeRoleToUserCache(username, newRole);
   }
 
   private ExecutionResponse getServerEvents(final String command) {
@@ -185,13 +231,15 @@ public class PostServerCommandHandler extends AbstractHandler {
     final ArcadeDBServer server = httpServer.getServer();
     server.getServerMetrics().meter("http.get-server-events").hit();
 
-    final JSONArray events = fileName.isEmpty() ? server.getEventLog().getCurrentEvents() : server.getEventLog().getEvents(fileName);
+    final JSONArray events = fileName.isEmpty() ? server.getEventLog().getCurrentEvents()
+        : server.getEventLog().getEvents(fileName);
     final JSONArray files = server.getEventLog().getFiles();
 
     return new ExecutionResponse(200, "{ \"result\" : { \"events\": " + events + ", \"files\": " + files + " } }");
   }
 
   private ExecutionResponse listDatabases(final ServerSecurityUser user) {
+    // All users can list databases, no check is needed
     final ArcadeDBServer server = httpServer.getServer();
     server.getServerMetrics().meter("http.list-databases").hit();
 
@@ -216,17 +264,26 @@ public class PostServerCommandHandler extends AbstractHandler {
     database.command("sql", "align database");
   }
 
-  private void dropDatabase(final String command) {
+  private void dropDatabase(final String command, final ServerSecurityUser user) {
     final String databaseName = command.substring("drop database ".length()).trim();
     if (databaseName.isEmpty())
       throw new IllegalArgumentException("Database name empty");
 
-    final Database database = httpServer.getServer().getDatabase(databaseName);
+    // check if user has create database role, or is root
+    var anyRequiredRoles = List.of(ServerAdminRole.DROP_DATABASE, ServerAdminRole.ALL);
+    if (httpServer.getServer().getSecurity().checkUserHasAnyServerAdminRole(user, anyRequiredRoles)) {
+      final Database database = httpServer.getServer().getDatabase(databaseName);
 
-    httpServer.getServer().getServerMetrics().meter("http.drop-database").hit();
+      httpServer.getServer().getServerMetrics().meter("http.drop-database").hit();
 
-    ((DatabaseInternal) database).getEmbedded().drop();
-    httpServer.getServer().removeDatabase(database.getName());
+      ((DatabaseInternal) database).getEmbedded().drop();
+      httpServer.getServer().removeDatabase(database.getName());
+
+      /**
+       * TODO in keycloak:
+       * 1. delete all roles for the removed database
+       */
+    }
   }
 
   private void closeDatabase(final String command) {
@@ -252,23 +309,27 @@ public class PostServerCommandHandler extends AbstractHandler {
   }
 
   private void createUser(final String command) {
-    final String payload = command.substring("create user ".length()).trim();
-    final JSONObject json = new JSONObject(payload);
+    // final String payload = command.substring("create user ".length()).trim();
+    // final JSONObject json = new JSONObject(payload);
 
-    if (!json.has("name"))
-      throw new IllegalArgumentException("User name is null");
+    // if (!json.has("name"))
+    // throw new IllegalArgumentException("User name is null");
 
-    final String userPassword = json.getString("password");
-    if (userPassword.length() < 4)
-      throw new ServerSecurityException("User password must be 5 minimum characters");
-    if (userPassword.length() > 256)
-      throw new ServerSecurityException("User password cannot be longer than 256 characters");
+    // final String userPassword = json.getString("password");
+    // if (userPassword.length() < 4)
+    // throw new ServerSecurityException("User password must be 5 minimum
+    // characters");
+    // if (userPassword.length() > 256)
+    // throw new ServerSecurityException("User password cannot be longer than 256
+    // characters");
 
-    json.put("password", httpServer.getServer().getSecurity().encodePassword(userPassword));
+    // json.put("password",
+    // httpServer.getServer().getSecurity().encodePassword(userPassword));
 
-    httpServer.getServer().getServerMetrics().meter("http.create-user").hit();
+    // httpServer.getServer().getServerMetrics().meter("http.create-user").hit();
 
-    httpServer.getServer().getSecurity().createUser(json);
+    // httpServer.getServer().getSecurity().createUser(json);
+    throw new RuntimeException("Please create new user through keycloak");
   }
 
   private void dropUser(final String command) {
@@ -287,7 +348,8 @@ public class PostServerCommandHandler extends AbstractHandler {
     final HAServer ha = httpServer.getServer().getHA();
     if (ha != null && !ha.isLeader())
       // NOT THE LEADER
-      throw new ServerIsNotTheLeaderException("Creation of database can be executed only on the leader server", ha.getLeaderName());
+      throw new ServerIsNotTheLeaderException("Creation of database can be executed only on the leader server",
+          ha.getLeaderName());
   }
 
   private HAServer getHA() {
